requirements/testing.txt
WebTest
django-webtest
factory-boy
pyquery
freezegun
